+++ schemes/expenses.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel, Field
 
 
 class CreateExpense(BaseModel):
-    money: Decimal = Field(..., ge=0)
+    money: float = Field(..., ge=0)
     currency_id: int
     source: str
     source_id: int
@@ -14,7 +14,7 @@ class CreateExpense(BaseModel):
 
 class UpdateExpense(BaseModel):
     id: int
-    money: Decimal = Field(..., ge=0)
+    money: float = Field(..., ge=0)
     currency_id: int
     source: str
     source_id: int
